refactor(library): Extract component to represent Greater London Assembly elections
