+++ uk_election_timetables/elections/__init__.py
@@ -1,2 +1,3 @@
 from uk_election_timetables.elections.scottish_parliament import *
 from uk_election_timetables.elections.senedd_cymru import *
+from uk_election_timetables.elections.greater_london_assembly import *
